WIP wasp collisions
Trying to change wasp collisions so that they passthrough picnic items
when targeting player (so picnic items do not act as shields to player).

Think I have the code correct but it's not working! Looks like a unity
bug to me - if you change an objects layer at runtime, the object
evaluates the collision matrix using it's original layer number, not the
one it's been moved to. Lame.