WIP edible food between rounds
'Set up phase' logic adjusted to include a 5 second period before the
food is switched out for the next round, in which the food from the now
finished round can be eaten. Food has a collider to detect interaction
with the player, and when the collider is triggered, the "food is
consumable" phase timer jumps to 0 and the next food item is brought
out.