Rain system WIP
Cloud and rain system started. Rain objects spawn in storm/cloud object,
and fall with gravity until a randomly generated splash coordinate on y
is reached, then the sprite animates a splash. Gravity and velocity are
disabled for splash, and then rain object is destroyed after a moment.

Draw order for rain objects is determined by the splash coordinate,
allowing them to draw on top of player while higher up the screen, but
due to splash below player.

Because rain objects are child objects in the storm/cloud, the cloud can
roll in from a direction and the rain will generate as a front that
follows the cloud, ie, splash coordinate will generate as a roughly set
distance below cloud at time of creation.