Fixing healthbar visibility through storm cloud
The healthbar GUITexture was drawing prominently through the cloud. Now
it fades to alpha=0, as the raincloud fades to alpha=255. The 'special
alpha' variable is stored in the _GM setup script, and set by the cloud
as it changes, and read by health bar per update(). The healthbar
variable = 1- var held by GM. Default value in GM is -254 so that the
healthbar gets an alpha of 1, until the cloud starts to fade. The cloud
sets the GM var at this point to whatever as it goes totally opaque, and
the hbs then read 1-that to fade out simultaneously.

Not a nice implementation at all, but with all the healthbar cloning
business, this was how I eventually got it to work.