Virtual joystick for touch screen (android)
Got a virtual joystick implemented, based on the 'single joystick'
standard assets (mobile) prefab. I've replaced the javascript script
with a c# version found online.

Plenty of work to do such as proper platform detection, and the
associated UI changes. But at the moment the game is basically playable
on android.